% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{add_eucl_dist}
\alias{add_eucl_dist}
\title{Add Miminal euclidean distance to grid data}
\usage{
add_eucl_dist(data)
}
\arguments{
\item{data}{(list of 5) 1. grid, 2. stations
3. selected_vars 4. weights 5. name of pollutant}
}
\value{
data (list of 5) 1. grid, 2. stations with minimal euclidean distances
and 3. selected_vars 4. weights 5. name of pollutant
}
\description{
Add Miminal euclidean distance to grid data
}
