% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{add_var_info}
\alias{add_var_info}
\title{Adds information about the feature importance of the Variables}
\usage{
add_var_info(data_raw, var_list)
}
\arguments{
\item{data_raw}{(list of 2) 1. grid and 2. stations}

\item{var_list}{(list of 2) 1. selected_vars 2. weights}
}
\value{
data_raw_info (list of 5) 1.grid 2.stations
3. selected_vars 4. weights 5. name of pollutant
}
\description{
This function adds a list with a vector containing the names of the needed
variables and a vector containing a weighting factor (e.g. derived from the
variable importance) for each variable
}
