% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_hotspot_utils.R
\name{comma_separated_chunks}
\alias{comma_separated_chunks}
\title{Divide a vector into chunks and format them as comma-separated strings}
\usage{
comma_separated_chunks(vec, chunk_length)
}
\arguments{
\item{vec}{a vector, e.g., of x coordinates}

\item{chunk_length}{(integer) maximal chunk size; Inf to return exactly one chunk}
}
\value{
list of strings like, e.g., "3, 4, 5"
}
\description{
The chunks can be used for example in SQL queries
}
