% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{compute_dist_mat}
\alias{compute_dist_mat}
\title{Compute Distance Matrix}
\usage{
compute_dist_mat(data)
}
\arguments{
\item{data}{(list of 5) 1. grid, 2. stations 3. selected_vars 4. weights
5. name of pollutant}
}
\value{
dist_mat of Size: rows of grid x rows of stations
}
\description{
This Function computes a distance matrix between two data.frames with a
euclidean distance measure
}
