% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotting.R
\name{create_all_pollutants_map}
\alias{create_all_pollutants_map}
\title{Create hotspot maps for all three pollutants}
\usage{
create_all_pollutants_map(
  hotspot_data,
  metric,
  frac_render,
  type = "layered",
  save = FALSE
)
}
\arguments{
\item{hotspot_data}{(data.frame) dataframe for a certain pollutant}

\item{metric}{(string) ("avg", "quantile_90", or "max"}

\item{frac_render}{(numeric) (numeric) What share of all data should be
rendered? Top frac_render*100\% of data will be kept. Defaults to 10\% (=0.1)}

\item{type}{(string) "layered" or "synced"}

\item{save}{(bool) save file?}
}
\description{
Create hotspot maps for all three pollutants
}
