% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotting.R
\name{create_cells_map}
\alias{create_cells_map}
\title{Plots a map of filtered raster cells for a certain layer}
\usage{
create_cells_map(
  hotspot_data,
  column,
  extra_layer_info = "",
  frac_render = 0.1,
  hide_layer = FALSE
)
}
\arguments{
\item{hotspot_data}{(data.frame) dataframe for a certain pollutant}

\item{column}{(string) column of dataframe to plot in map as layer}

\item{extra_layer_info}{(string) additional string to name layer - will be
used as a prefix for default layer name. Defaults to empty string.}

\item{frac_render}{(numeric) (numeric) What share of all data should be
rendered? Top frac_render*100\% of data will be kept. Defaults to 10\% (=0.1)}

\item{hide_layer}{(bool) Defaults to FALSE - should layer be hidden?
(= unticked)}
}
\value{
map with stations (mapView Object)
}
\description{
Plots a map of filtered raster cells for a certain layer
}
