% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_hotspot_utils.R
\name{create_cells_sf}
\alias{create_cells_sf}
\title{Turn hotspot data into a filtered, spatial sf object}
\usage{
create_cells_sf(hotspot_data, metric, frac)
}
\arguments{
\item{hotspot_data}{(data.frame) dataframe for one pollutant with}

\item{metric}{(string) filter on this metric (must be a column in
hotspot_data)}

\item{frac}{(numeric) What share of all data should be filtered? Top
frac_render*100 \% of the data will be kept.}
}
\value{
data.frame with cols x, y, avg, max, quantile_90
}
\description{
Turn hotspot data into a filtered, spatial sf object
}
