% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotting.R
\name{create_stations_map}
\alias{create_stations_map}
\title{Plots a Map of Berlin with all 18 stations}
\usage{
create_stations_map()
}
\value{
map with stations (mapView Object)
}
\description{
Plots a Map of Berlin with all 18 stations
}
