% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_hotspot_utils.R
\name{create_stations_sf}
\alias{create_stations_sf}
\title{Get Coordinates of measuring stations and turn them into a sf object (points)}
\usage{
create_stations_sf()
}
\value{
sf object with points
}
\description{
Get Coordinates of measuring stations and turn them into a sf object (points)
}
