% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{drop_gridpoints_eq_to_station}
\alias{drop_gridpoints_eq_to_station}
\title{Drop all grid points that are equal to a station}
\usage{
drop_gridpoints_eq_to_station(data)
}
\arguments{
\item{data}{(list of 5) 1. grid, 2. stations 3. selected_vars 4. weights
5. name of pollutant}
}
\value{
dataFiltered (list of 5) 1. grid, 2. stations
3. selected_vars 4. weights 5. name of pollutant
}
\description{
This Function remove all grid points that are equal to a station,
but not the stations themself.
}
