% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{filter_distances}
\alias{filter_distances}
\title{Filter Distances}
\usage{
filter_distances(data, lower_bound, upper_bound = 1)
}
\arguments{
\item{data}{(list of 5) 1. grid, 2. stations 3. selected_vars 4. weights
5. name of pollutant}

\item{lower_bound}{(numeric) lower quantile bound in percentage}

\item{upper_bound}{(numeric) upper quantile bound in percentage}
}
\value{
filtered data (list of 5) 1. grid, 2. stations 3. selected_vars
4. weights 5. name of pollutant
}
\description{
This Function filter
}
