% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotting.R
\name{generate_heat_map}
\alias{generate_heat_map}
\title{Plots a Heatmap of for FAirQ Novelty Detection}
\usage{
generate_heat_map(data, zcol = "min_dist")
}
\arguments{
\item{data}{(list of 5) 1. grid, 2. stations 3. selected_vars 4. weights
5. name of pollutant}

\item{zcol}{(string) name of column, which should be
displayed on the heat map}
}
\value{
(mapView Object) heat map of mimimal distances
}
\description{
This Functions Plots a Heatmap
}
