% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotting.R
\name{generate_station_map}
\alias{generate_station_map}
\title{Plots a Map of Berlin with all 18 stations}
\usage{
generate_station_map(data)
}
\arguments{
\item{data}{(list of data.frames) 1. Grid with x,y Coordinates and a Distance
2.Station}
}
\value{
(mapView Object) map with stations
}
\description{
Plots a Map of Berlin with all 18 stations
}
