% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_hotspot_utils.R
\name{get_all_x_coords}
\alias{get_all_x_coords}
\title{Get all x coordinates of Berlin city raster as numeric vector}
\usage{
get_all_x_coords()
}
\value{
numeric vector containing all x coords sorted
}
\description{
Get all x coordinates of Berlin city raster as numeric vector
}
