% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_hotspot_utils.R
\name{get_hotspot_data}
\alias{get_hotspot_data}
\title{Gets hotspot data for all x coords}
\usage{
get_hotspot_data(pollutant, x_coords_chunk_length)
}
\arguments{
\item{pollutant}{(string) "no2", "pm10" or "pm25}

\item{x_coords_chunk_length}{(integer) How many x coordinates should be
queried at once?}
}
\value{
data.frame with cols x, y, avg, max, quantile_90
}
\description{
Gets hotspot data for all x coords
}
