% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_hotspot_utils.R
\name{get_partial_hotspot_data}
\alias{get_partial_hotspot_data}
\title{Gets hotspot data for one chunk of x coords}
\usage{
get_partial_hotspot_data(relevant_model_id, x_coord_chunk)
}
\arguments{
\item{relevant_model_id}{(integer) Output of get_relevant_model_id, refers
to one pollutant implicitly}

\item{x_coord_chunk}{(vector of numerics) Vector of x coordinates to query}
}
\value{
data.frame with cols x, y, avg, max, quantile_90
}
\description{
Gets hotspot data for one chunk of x coords
}
