% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_hotspot_utils.R
\name{get_relevant_model_id}
\alias{get_relevant_model_id}
\title{Gets relevant model id for hotspot full model}
\usage{
get_relevant_model_id(pollutant)
}
\arguments{
\item{pollutant}{(string) "no2", "pm10" or "pm25"}
}
\value{
scalar (numeric)
}
\description{
Gets relevant model id for hotspot full model
}
