% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{join_raw_data}
\alias{join_raw_data}
\title{Join data with min distances with original raw data}
\usage{
join_raw_data(data, data_raw)
}
\arguments{
\item{data}{(list of 5) 1. grid, 2. stations
3. selected_vars 4. weights 5. name of pollutant}

\item{data_raw}{(list of 2) 1. grid, 2. stations}
}
\description{
This function joins the computed min_distances in data with the
original untransformed information data_raw
}
