% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_load.R
\name{load_data}
\alias{load_data}
\title{Load all data for novelty detection}
\usage{
load_data()
}
\value{
list of data.frames
}
\description{
Loads all necessary data for novelty detection.
Save data in a Rdata file
and return data as list of data frames
}
