% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{multiply_by_weight}
\alias{multiply_by_weight}
\title{Multiply variables by weighting factor}
\usage{
multiply_by_weight(data)
}
\arguments{
\item{data}{(list of 4) 1. grid, 2. stations 3. selected_vars 4. weights
5. name of pollutant}
}
\value{
(list of 5) 1. grid, 2. stations with weighting factor
3. selected_vars 4. weights 5. name of pollutant
}
\description{
This function multiplies each variable by a given weighting factor
(e.g. derived by variable importance)
}
