% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotting.R
\name{save_map}
\alias{save_map}
\title{Save mapview Object as html file}
\usage{
save_map(map, path = "./inst/reports/")
}
\arguments{
\item{map}{(mapview object) the map to be saved}

\item{path}{(string) output path where html}
}
\description{
This function saves a mapview object map as html file under the
same name as the name of the map variable.
}
