% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{select_var_from_data}
\alias{select_var_from_data}
\title{Select variables from grid}
\usage{
select_var_from_data(data)
}
\arguments{
\item{data}{(list of 5) 1. grid, 2. stations 3. selected_vars 4. weights
5. name of pollutant}
}
\value{
data (list of 5) 1. grid, 2. stations with selected
variables and 3. selected_vars 4. weights 5. name of pollutant
}
\description{
This function selects the subset of variables stored in selected_vars in the
grid and stations dataframe
}
