% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{standardization}
\alias{standardization}
\title{Data Standardization}
\usage{
standardization(data, quantile = 0.95, cutoff = 1)
}
\arguments{
\item{data}{(list of 5) 1. grid, 2. stations 3. selected_vars 4. weights
5. name of pollutant}

\item{quantile}{(numeric) quantile for standardization}

\item{cutoff}{(numeric) upper cutoff value}
}
\value{
(list of 5) 1. grid, 2. stations with robust standardized_grid
3. selected_vars 4. weights 5. name of pollutant
}
\description{
This functions computes a standardization of the, using a given quantile
}
